pyproject.toml
[tool.poetry]
name = "sprint-4"
version = "0.1.0"
description = ""
authors = ["Alexandra Nikulina <alxndra.nikulina@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.3.1"
allure-pytest = "^2.13.1"
selenium = "^4.9.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
